# Conflicts:
#	pom.xml
#	src/main/java/sarik/dev/foodwaveproject/repository/OrderItemRepository.java

+++ src/main/java/sarik/dev/foodwaveproject/entity/CartItem.java
@@ -12,20 +12,21 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class CartItem {
 
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long cartItemId;
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long cartItemId;
 
-	@ManyToOne
-	@JoinColumn(name = "cart_id")
-	private Cart cart;
+    @ManyToOne
+    @JoinColumn(name = "cart_id")
+    private Cart cart;
 
-	@ManyToOne
-	@JoinColumn(name = "product_id")
-	private Product product;
+    @ManyToOne
+    @JoinColumn(name = "product_id")
+    private Product product;
 
-	private Integer quantity;
-	private double discount;
-	private double productPrice;
+
+    private Integer quantity;
+    private long discount;
+    private long productPrice;
 
 }
